+++ transpiler/src/test/java/org/jsweet/test/transpiler/ApiTests.java
@@ -277,7 +277,7 @@ public class ApiTests extends AbstractTest {
 		eval(ModuleKind.none, (logHandler, r) -> {
 			logHandler.assertNoProblems();
 			String localeString = r.get("localeString");
-			List<String> expected = asList("2020-1-1 01:00:00", "1/1/2020, 12:00:00 AM");
+			List<String> expected = asList("2020-1-1 01:00:00", "1/1/2020, 12:00:00 AM", "01/01/2020 à 01:00:00");
 			assertTrue("date is not ok: " + localeString + " is not one of " + expected,
 					expected.contains(localeString));
 		}, getSourceFile(Dates.class));
